refactor: extract campsnap browser helpers
